Add an x hotkey for fake fleet orders.
If you select an enemy fleet you can now press 'x' to give it
new orders in order to explore scenarios in the battle calculator.

If you select an enemy star, a fake fleet will be created there
for you to direct.

Reload when you're done to reset the game data.