Make a distinction between who you're controlling and you.
When using 'x' to control enemy fleets, note that in the lower
right of the HUD with the text 'controlling <alias>'; and if
you're viewing as someone else, don't lose that info. So you can
view as player a and control player b and see A controlling B as
the indication of the unreal state you are in.