Finally get all cases for war and peace.
In the formal alliances report, capture agreed peace, observed
peace, declared war, and war, so that the battle calc will respect
the war state when returned to after a period of peace. This gets
all the cases: there is no need to 'observe war' because war is
the default and must be formally declared after peace.

The only edge case is if two players are at peace and you never
observe it: you cannot then know when they were at peace if they
first declare war.