Add screenshot functionality for users of imgbb.com.
You can create an imgbb.com account to host screenshots and get
an API key from them. Putting your API key into NPA's settings
allows the extension to upload your current map view as a screenshot
for sharing with your allies.

This involved a fairly big overhaul of the settings to have types
and allowed values and a generic form for setting all of them. It
mostly works; it struggles with editing the text fields but
otherwise seems OK. It's good enough to be getting on with even if
not ideal.