Couldn't remember why I changed clipboard setting.
Originally, if you ever viewed a report using the UI, it would
be set as the 'lastClip' and put on the clipboard. I couldn't
remember why I changed this to not copy by default, and it
wreaks havoc with async report generation. For now, make things
go to both places again so that I can see what the problem was.

Also gut the example test for hotkey.ts because I couldn't be
bothered how to deal with the navigator call and there are now
better (real) test examples in the codebase, for example patch.ts
and patch.spec.ts.