Fix economists report based on @Skilly146's bug report.
Most embarrassingly, it looks like when I refactored the code
to include @BelSon's FR for the buy vs trade for terra I broke
the output table and didn't include terra in the analysis of
what you can buy after production, making banking always look
better.

Banking is still better a surprising amount of the time, but not
always!