Change approach for alliance definition/coloring.
Instead of deciding what colour someone is based on what alliance
they are in, set it up so that the user will be able to configure
colours and shapes for every player as an orthogonal feature.

Then, we can provide alliance modes that decide who's allied based
on them sharing either a common symbol or a common colour at the
user's option, and the user doesn't have to use alliance inferences
at all if they don't want to. So the alliance mode will be "group
by colour", "group by symbol", or "don't assume alliances", and the
default will be "group by colour". Setting two players to the same
colour will then make them allied as far as the rest of the reports
are concerned, starting first and foremost with the empires report.