Move custom colours into a setting, fix route enemy help.
Move the custom colours into their own setting to prepare for
allowing the user to configure colours for every player in the
game. While I was at it, noticed the help being broken for
route enemy (as well as the control button) and fixed it to
have a better help string and a reasonable button name.