Remove home planets report at startup.
In order to avoid always clobbering the user's clipboard with a
report they are unlikely to need now that the UI has autocomplete,
don't put the home planets on the clipboard at startup anymore.

This also might have been the source of the zillions of "Document
is not focused" errors.