Redo the activity report.
Redo the activity report to make it easier to spot AFK players and
easier to consume in a tabular format. This doesn't give quite all
the data that are there, only the last 5 actions the player took,
but I think this is 90% of the utility of this form of the report.
Separately an activity histogram per player might be really nice,
but I am not sure if that should be in this report or a new report
of its own.