Fix an egregious bug in the activity report.
Though the code purportedly checked for tick changes for industry
and science based activity, in fact the dereferenced field was
always undefined so all star captures were showing up as manual
activity on behalf of the player.

Fix it to use the actual tick number and report activity
conservatively as it should.