Add counters to track API calls.
Suspicious that v2.2.8 started unduly hammering Jay's np servers
with requests, I have added the ability to log arbitrary counts
and am now logging counts of all post requests coming from the
extension so that if there is something spinning out of control
it can be seen in the counters. Counters are per version number.