Make caching of scan data more conservative.
In the original code, it didn't wait for the API data to be written
to the cache before returning it, opening up the possibility that
repeated calls would hit the real API endpoint multiple times
needlessly. Though it introduces latency in the client, wait for
this to get written to the database to make that scenario impossible.

Also, refreshScanData was willing to refresh multiple times as it is
async and could be called repeatedly triggering multiple API calls
for the same scan. The actual effect of this was two API calls per
key at load time, one of which was needless. Change it to refuse to
refresh again if it has been called in the last five minutes.

Also, don't hook on_full_universe for refreshScanData as it is
already wired to refresh_interface, so only one was needed.

All of this in the hope of reducing API calls to the server, but
none of these fixes look high impact to me. Still better than it
was even if it doesn't fix the real problem.