In proteus, industry and science get destroyed on conquest.
The combat calculator was still granting partial production
and ongoing production to the next owner of a star after
combat, but combat eliminates all industry on the star and
so it is really unknown what production will be after losing.

It starts at 0 and the other player can of course upgrade if
the are online/awake to do so. Update the calculator to assume
no upgrades and therefore no production at the destination for
proteus.