Handle undefined response to game messages.
It looks like occasionally the response that comes back doesn't
contain game messages - perhaps there's an auth problem or an
error in the underlying game - so check for and handle this.

Also make another attempt to find a message in unhandled promise
rejections which often have no message and no stack trace.