Add in more logging and code from v2.2.41.
Version v2.2.48 doesn't reproduce the request messages bug after
a week of waiting, yet seems to contain all the code that was in
the original faulty diff between v2.2.7 and v2.2.8. It was missing
one block from v2.2.41, the last version where the bug was observed.

Deploy this with the safety in place so that it shouldn't actually
hammer the real server, to see if the bug is still reproducing in
the wild.