Modify activity report to not consider tick 0 updates.
A player will go AFK if they didn't make any updates after
the start of the game, but activity was crediting pre-game
activity to tick #1 if the player was active multiple times
in the pre-game interval. Squish all that together so that
they show up as possibly AFK instead.

Also, make the activity report only run up to the current
tick on display, rather than the current real tick, so that
you can scroll back through activity with the time machine.