Change ownership report to show only the last cycle.
Make the ownership report show the last 24h (or 20h, or whatever
depending on cycle length) so you can scroll through it with the
time machine and just see current changes when you pull it up.

Unsure of the goodness of this change as it works worse with
filtering but better for interactive exploring. Might make it
show full history if you apply a filter for the best of both
worlds.