Add the new communal economy report.
The economists would like to convince your alliance that since money
transfers are free, it's better to build the cheapest economy
communally across the alliance and split the cash after production
rather than have each member ramp their economy individually. In
this way, the exponential growth will ensure the alliance is rolling
in it sooner.

Of course, this just requires a little trust.

The new economists report tells you how to rebalance cash for econ
buys and outputs a table of who buys econ on what stars.

It adds three new [[ ]] expansions, one to show a player's cash,
one to enabling transferring them the right amount of cash, and one
to enable a bulk upgrade button for upgrading stars.

It adds a new feature to merging scan data where player data are
now merged into the main view and so it is easier to get information
about the other player's research and cash based on what is currently
merged.

It adds a new feature to combatOutcomes where the caller can pass
in an object to get back the star states after combat, so that the
economists will decline to propose that you buy economy on a star
that you can already see you will lose before the cycle. If you
lose it and gain it in the combat sequence, they will still wrongly
recommend it.

Written in a sequence of pair programming sessions with Rob! Thanks!