Fix recolor to white to not get persisted by accident.
In the recent change to persist colour changes all the time, I
broke whiteout because it now gets set as the user's custom
color instead of being tracked as a separate temporary mode.

Fix it to not record the white colour as the custom colour
so that you can toggle it on and off as usual.