Always watch for onServerResponse, and load later.
Originally the if statement that checked for whether or not
it was the right timing to initialize NPA was the very last
thing in the file, but more code got added under it. Move the
test for early load to the bottom so that hotkeys are initialized
for all the reports and in general so that the load order is as
expected, meaning the entirety of NeptunesPrideAgent() has been
evaluated so that everything is defined.

As part of that, always hook onServerResponse which gets hooked
most of the time in practice, so that server events do always
initialize NPA. Probably some of that logic could be simplified,
as there is really no need to respond to anything other than
full universe messages.

As a result the menu always displays the hotkeys instead of
sometimes failing to show them.