Time machine fixes for NP4.
Make sure scanInfo contains the correct  value for puid now that
the underlying scan data has a new name for it (playerUid instead
of player_uid) and fix up access of productionTicks to point at
the new config in the galaxy object instead of the old (and
now empty) config on universe.

Makes the ownership and trading reports work again.