Fix generals bug reported by StSeanSpicer on discord.
Generals was not recording losses on the star due to a typo where
a - was replaced by an = in the calculation. Not only that, but
the wrong data would then get cached so even with this bug fix
in place clients will still get the wrong numbers until they play
new games / fight new fights.