Fix map overlays to work with the new style map code.
Jay pushed a change that rewrote the map as a class, so all
overrides were broken and map drawing was wrong. Fix the overrides
to bind the correct functions so things seem to mostly work.

This change got mixed in by accident with a change to fix trade
reporting to not report trades on tech that is newly visible due
to an FA being accepted.