Switch from webpack to esbuild.
In the hopes that it'll be easier to figure out how to do
non-bundled imports of the new NP4-style classes, switch
from webpack to the much faster esbuild tool.

In this version, there is a problem where it constantly
rebuilds twice because of the way I compute the version.js
file, but since each build takes about 100ms and webpack
took about 7000ms, we'll call the 35x speed improvement
good for now.