Fix game_diplomacy indexing for new message comments.
New message comments are no longer identified by knowing the
comment count and no longer have unique keys. Instead, all
comments are given the key of the thread they are in.

Change the indexeddb representation to allow duplicate keys
for comment databases and the surrounding code to no longer
rely on knowing the comment_count either to differentiate
cases or to fetch the minimal number of comments. Also change
the regexp for recognizing API keys to match NP4 keys as well
as NP2 keys, so that when you switch browsers you can find
the API keys that had been mailed to you in the key report.