Make route planner show cumulative time.
On each route planner arrow show the cumulative time for
the proposed path. Also add an 'L' hotkey to lock the route
planner display so that the user can easily implement the
suggested routes, with reference to the times so that it's
clear how long the ships will take to reach the selected
front.

Remove a noisy CALC console.log from timetravel.