Fix full cycle future time machine.
The new galaxy state wasn't being updated on each iteration of
the loop properly, so the future time machine was showing bad
results when jumping entire cycles at once. Fix it to copy the
state on each iteration and update it appropriately so that
battle calc problems don't occur, and now it gets the same
results as single stepping the time machine to the same tick.