Use a BSP tree for scan range visibility for speed.
Change all report screens to include timing at the bottom to make
it easier to spot performance problems and improvements.

Remove the bespoke timing from the recently optimized activity
report.

Reviewed all screens for timing in a 64p game. The trading report
was running pretty slowly due to the way it calculated scan range
visibility; redo it to be based on a bsp tree, speeding it up from
about 2-2.5s to 0.5s in the case I tested with.