Change the filter strategy to whole stanzas.
It used to be a filter had to match a single line of a stanza
for that group to be selected by the filter, but this meant that
in "All Combats" there was no way to find all combats that affect
a certain player; only those where that player was specifically
defending.

Changing it to match against the entire stanza at once means you
can easily review all combats that affect someone (e.g. yourself)
by typing their alias into the filter box ... and as far as I can
see doesn't hurt any of the other filtering scenarios.