Reorder variable declarations to avoid ReferenceError.
Reported on github, it seems that sometimes the initialization order
is more strict than others. The technical rules of let and const
declarations are that this should never work, but it often did work
for me. This change should fix it in all cases for these variables:
time travel ticks, setPlayerColor, and colorMap.