First pass of historical time machine rewrite.
Switch over to a new, hopefully simpler and bug-free scheme
of saving the raw server side diff blocks and processing them
in memory into the old data structure.

Right now this does read all the blocks again and again, but
other than that seems at least as efficient as before, and on
the upside, also works.