Fix the colours flashing when the galaxy updates.
The expandPlayerData method of the base game resets the colours to
system colours, and the best way I could find to deal with it was
to fix it up based on what the player colorus just were before any
drawing can happen.

So the fleet lines no longer flash back to the original colour and
advancing rapidly through time machine looks good.

However, historical time machine has an unrelated bug that shows
up when testing this where it encounters a bad back patch. The
rewrite of npaserver still looks necessary, though the time machine
is working better than it has in a while.