Fix a bug in the route planner that BelSon found.
BelSon pointed out that the route planner wasn't finding shortest
paths to all the leaves and after confusing myself as to whether
I'd implemented an MST Prim-style or Dijkstra-style, I realized
the code was in fact meant to find a spanning tree of all shortest
paths and was failing to do so.

The problem was not computing the time required between stars in
complete ticks, but rather using and summing fractional ticks, which
ignores the rounding up overhead of each stop and considers paths
to be shorter than they are in actual gameplay. This error was also
reflected on the display which converted to rounded tick numbers for
the total distance rather than summing the distances down the edges
of the path.

In the end a trivial change to use integers for the individual edges
corrects the computation and gives the desired (and now also expected;)
result.