Timeslice time machine data processing.
In big games with FAs and hundreds of ticks, it takes multiple
seconds for the new patch code to build up and validate the time
machine data structures. Slice this up with short delays so that
the main thread can still do stuff for the user, reducing load
time jank.