Fix speed calculation for NP4.
In NP2 there was no way to reference the source star and destination
star to accurately calculate speeds in future time machine given
warp gate changes. But in NP4 this is fixed by having each fleet
record lsuid - the id of the source star - as well as knowing its
destination. This means that if both stars are visible to the player,
a more accurate speed can be calculated for future time machine
that takes into account gates constructed or destroyed during the
same tick that the user is in when exploring future time scenarios.

Use this info for more accurate future time, resolving the bug pointed
out by KoDBigMatt on discord.