Add auto route to fleets.
If you trigger fleet waypoint editing on a fleet that has no orders,
and you have the route planner on and locked to a destination,
pre-populate the fleet orders with the path to the destination with
default orders until arrival, drop all at arrival, and drop all for
the return flight. Show the route editing screen to let the user
confirm the route, or clear it if that's not what they wanted to
do.