Fix autoruler for NP4 speed rules.
The autoruler was using a warp gate factor of 3 and applying
an odd proteus fix for wormholes and this was making it get
wrong results in NP4. Change it for NP4 range is speed rules,
and make it draw when a little more zoomed out so that you can
see the situation in huge maps with high range factors.

Disable drawing scanning ETA information at that scale though, since
you can't even see the fleets being referred to.