Update the new generals report to include time.
The generals now also report time to finish the current level
of each tech you might choose, as well as the time required
for the next level. This might help with tradeoffs involving
strengths and weaknesses in research, or where a much lower
level of a neglected tech is very easy to achieve now.