Fix two independent bugs, on in format, one in polititcalMap.
The Crux.format override would get badly broken if there was the
start of a substitution string nad no close ]] to terminate it.
Move the early exit up to avoid the exception.

politicalMap's cache wasn't being refreshed on loadGalaxy, so viewing
as another API key would have an offset display. Fix it to refresh in
this case.