Fix up a long-standing bug with filtering.
The reportHook was getting called over and over again for each filtering
keystroke because the global event handler was getting wired up repeatedly
to each new instance of the screen, which happens every time the user
uses ` or every time the 30s mark passes with NPA's UI displayed and
the timer ticking.

Hack around this by making only one global "exec_report" handler, and
having it call a singleton handler that gets replaced to point at the
current screen's handler, so that it is called exactly once per keystroke.

Shorten the timeout for typing to make the screen feel a little more
responsive. It is still the case that it feels kind of slow for slow
reports like trading, but nothing like before where it was potentially
multiple seconds.